PTX design fixes
