+++ src/features/ProtectTransaction/components/ProtectTxModalBackdrop.tsx
@@ -4,10 +4,11 @@ import styled from 'styled-components';
 
 const ProtectedTransactionModalBackdropLayout = styled.div`
   position: fixed;
+  height: 100vh;
+  width: 100vw;
+  z-index: 998;
   top: 0;
   left: 0;
-  right: 0;
-  bottom: 0;
   background: rgba(255, 255, 255, 0.5);
 `;
